Remove redux from Forgot Password screen
